Update the timer template to use the new timer UI
Change the POST request in the EnsembleTimerController to take the Ensemble ID as a Path parameter (variable) instead of a Query (request) parameter

+++ src/test/java/com/jitterted/mobreg/domain/EnsembleFactory.java
@@ -60,7 +60,7 @@ public class EnsembleFactory {
     }
 
     @NotNull
-    public static Ensemble withStartTimeNowAndIdOf(int id) {
+    public static Ensemble withStartTimeNowAndIdOf(long id) {
         Ensemble ensemble = new Ensemble("Starts Now", ZonedDateTime.now());
         ensemble.setId(EnsembleId.of(id));
         return ensemble;
